+++ src/SFA.DAS.EmployerApprenticeshipsService.Database/EnglishFraction.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[EnglishFraction]
 (
-	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
+	[Id] BIGINT NOT NULL PRIMARY KEY IDENTITY, 
     [DateCalculated] DATETIME NULL, 
     [Amount] DECIMAL(18, 4) NULL, 
     [EmpRef] NCHAR(50) NULL
